Enum symbiosis_rust::Collection

pub enum Collection<'a, T: 'a> {
    Slice(&'a [T]),
    Vec(Vec<T>),
    Map(BTreeMap<String, T>),
}

Variants

Slice

Vec

Map

Methods

impl<'a, T> Collection<'a, T>

fn values<'b>(&'b self) -> Box<Iterator<Item=&T> + 'b>

Get an iterator for the values in the collection.

fn key_values(&self) -> KeyValues<'a, &T>

Get an iterator for the keys and values in the collection.