Trait symbiosis_rust::Collection

pub trait Collection<'a, T: 'a> {
    fn values(&'a self) -> Box<Iterator<Item=T> + 'a>;

    fn key_values(&'a self) -> KeyValues<'a, T> { ... }
}

A generic collection trait.

Required Methods

fn values(&'a self) -> Box<Iterator<Item=T> + 'a>

Get an iterator for the values in the collection.

Provided Methods

fn key_values(&'a self) -> KeyValues<'a, T>

Get an iterator for the keys and values in the collection.